refactor `tls::trusted`
